showing total items on database and memory

+++ app/src/main/java/com/oxeanbits/bigstorageapp/AppState.kt
@@ -2,7 +2,7 @@ package com.oxeanbits.bigstorageapp
 
 data class AppState (
     val items: List<Request> = ArrayList(),
-    val count: Int = 0,
+    val count: Long = 0,
     val page: Int = 1,
     val pageSize: Int = 100,
     val lowMemory: Boolean = false
